prevention/pyproject.toml
[tool.poetry]
name = "prevention"
version = "0.1.0"
description = ""
authors = ["k29laksh <klaxay0505@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "5.0.3"
aiohttp = "3.9.3"
marshmallow = "3.21.1"
aiosignal = "1.3.1"
annotated-types = "0.6.0"
anyio = "4.3.0"
asgiref = "3.8.1"
attrs = "23.2.0"
build = "1.2.1"
cachecontrol = "0.14.0"
certifi = "2024.2.2"
charset-normalizer = "3.3.2"
cleo = "2.1.0"
colorama = "0.4.6"
crashtest = "0.4.1"
dataclasses-json = "0.6.4"
distlib = "0.3.8"
distro = "1.9.0"
django-cors-headers = "4.3.1"
djangorestframework = "3.15.1"
djangorestframework-simplejwt = "5.3.1"
dulwich = "0.21.7"
fastjsonschema = "2.20.0"
filelock = "3.15.4"
frozenlist = "1.4.1"
greenlet = "3.0.3"
h11 = "0.14.0"
httpcore = "1.0.5"
httpx = "0.27.0"
idna = "3.6"
installer = "0.7.0"
jsonpatch = "1.33"
jsonpointer = "2.4"
keyring = "24.3.1"
langchain = "0.1.13"
langchain-community = "0.0.29"
langchain-core = "0.1.36"
langchain-text-splitters = "0.0.1"
langsmith = "0.1.37"
more-itertools = "10.3.0"
msgpack = "1.0.8"
multidict = "6.0.5"
mypy-extensions = "1.0.0"
numpy = "1.26.4"
openai = "1.14.3"
orjson = "3.10.0"
packaging = "23.2"
pexpect = "4.9.0"
pillow = "10.4.0"
pkginfo = "1.11.1"
platformdirs = "4.2.2"
poetry = "1.8.3"
poetry-core = "1.9.0"
poetry-plugin-export = "1.8.0"
psycopg2 = "2.9.9"
ptyprocess = "0.7.0"
pydantic = "2.6.4"
pydantic-core = "2.16.3"
pyjwt = "2.8.0"
pyproject-hooks = "1.1.0"
python-decouple = "3.8"
python-dotenv = "1.0.1"
pywin32-ctypes = "0.2.2"
pyyaml = "6.0.1"
rapidfuzz = "3.9.4"
requests = "2.31.0"
requests-toolbelt = "1.0.0"
shellingham = "1.5.4"
sniffio = "1.3.1"
sqlalchemy = "2.0.29"
sqlparse = "0.4.4"
tenacity = "8.2.3"
tomlkit = "0.13.0"
tqdm = "4.66.2"
trove-classifiers = "2024.7.2"
typing-inspect = "0.9.0"
typing-extensions = "4.10.0"
tzdata = "2024.1"
urllib3 = "2.2.1"
virtualenv = "20.26.3"
yarl = "1.9.4"
django-split-settings = "^1.3.2"
ruff = "^0.5.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
